service/requirements.txt
Django==3.1.7
uWSGI==2.0.19.1
psycopg2-binary==2.8.6
requests==2.25.1
django-crispy-forms==1.11.1
openpyxl==3.0.7
django-filter==2.4.0
